Move all scope hoisting transforms together
