+++ src/assets/JSAsset.js
@@ -12,7 +12,7 @@ const babel = require('../transforms/babel');
 const generate = require('babel-generator').default;
 const uglify = require('../transforms/uglify');
 const SourceMap = require('../SourceMap');
-const hoist = require('../visitors/hoist');
+const hoist = require('../scope-hoisting/hoist');
 
 const IMPORT_RE = /\b(?:import\b|export\b|require\s*\()/;
 const ENV_RE = /\b(?:process\.env)\b/;
